requirements.txt
streamlit
openai
PyPDF2
python-docx
pillow
pytesseract
python-dotenv
